backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.2
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
deprecation==2.1.0
et_xmlfile==2.0.0
fastapi==0.109.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-vision==3.5.0
googleapis-common-protos==1.72.0
gotrue
grpcio==1.76.0
grpcio-status==1.62.3
httpx>=0.28.1
httpcore>=1.0.9
idna==3.11
numpy==1.26.4
openai
openpyxl==3.1.2
packaging==25.0
pandas==2.2.3
pillow==12.0.0
postgrest
proto-plus==1.26.1
protobuf==4.25.8
pypdfium2==4.26.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2025.2
realtime
reportlab==4.0.9
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
storage3
StrEnum==0.4.15
supabase
supafunc
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.27.0
websockets
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
